+++ tests/OnRamp.Test/CodeGenConsoleTest.cs
@@ -15,7 +15,7 @@ namespace OnRamp.Test
         [Test]
         public async Task A100_HelpDefaultOptions()
         {
-            var c = CodeGenConsole.Create<CodeGenConsoleTest>();
+            var c = new CodeGenConsole();
             var r = await c.RunAsync("--help");
             Assert.AreEqual(0, r);
         }
@@ -23,7 +23,7 @@ namespace OnRamp.Test
         [Test]
         public async Task A110_HelpSupportedOptions()
         {
-            var c = CodeGenConsole.Create<CodeGenConsoleTest>(options: SupportedOptions.IsSimulation | SupportedOptions.ExpectNoChanges);
+            var c = new CodeGenConsole(options: SupportedOptions.IsSimulation | SupportedOptions.ExpectNoChanges);
             var r = await c.RunAsync("--help");
             Assert.AreEqual(0, r);
         }
@@ -31,7 +31,7 @@ namespace OnRamp.Test
         [Test]
         public async Task A120_InvalidOptionScript()
         {
-            var c = CodeGenConsole.Create<CodeGenConsoleTest>();
+            var c = new CodeGenConsole();
             var r = await c.RunAsync("-s");
             Assert.AreEqual(1, r);
         }
@@ -39,7 +39,7 @@ namespace OnRamp.Test
         [Test]
         public async Task A130_InvalidOptionConfig()
         {
-            var c = CodeGenConsole.Create<CodeGenConsoleTest>();
+            var c = new CodeGenConsole();
             var r = await c.RunAsync("-c");
             Assert.AreEqual(1, r);
         }
@@ -47,7 +47,7 @@ namespace OnRamp.Test
         [Test]
         public async Task A140_InvalidOptionDirectory()
         {
-            var c = CodeGenConsole.Create<CodeGenConsoleTest>();
+            var c = new CodeGenConsole();
             var r = await c.RunAsync("-d ../../Bad");
             Assert.AreEqual(1, r);
         }
@@ -55,7 +55,7 @@ namespace OnRamp.Test
         [Test]
         public async Task A150_InvalidOptionAssembly()
         {
-            var c = CodeGenConsole.Create<CodeGenConsoleTest>();
+            var c = new CodeGenConsole();
             var r = await c.RunAsync("-a NotExists");
             Assert.AreEqual(1, r);
         }
@@ -63,7 +63,7 @@ namespace OnRamp.Test
         [Test]
         public async Task A160_InvalidOptionConnectionString()
         {
-            var c = CodeGenConsole.Create<CodeGenConsoleTest>();
+            var c = new CodeGenConsole();
             var r = await c.RunAsync("-db");
             Assert.AreEqual(1, r);
         }
@@ -71,7 +71,7 @@ namespace OnRamp.Test
         [Test]
         public async Task A170_InvalidOptionScriptSupportedOptions()
         {
-            var c = CodeGenConsole.Create<CodeGenConsoleTest>(options: SupportedOptions.IsSimulation | SupportedOptions.ExpectNoChanges);
+            var c = new CodeGenConsole(options: SupportedOptions.IsSimulation | SupportedOptions.ExpectNoChanges);
             var r = await c.RunAsync("-c ValidEntity.yaml");
             Assert.AreEqual(1, r);
         }
@@ -80,7 +80,7 @@ namespace OnRamp.Test
         public async Task A200_CodeGenException()
         {
             var a = CodeGeneratorArgs.Create<CodeGeneratorTest>("ValidEntity.yaml");
-            var c = CodeGenConsole.Create<CodeGenConsoleTest>(a);
+            var c = new CodeGenConsole(a);
             var r = await c.RunAsync();
             Assert.AreEqual(2, r);
         }
@@ -92,7 +92,7 @@ namespace OnRamp.Test
                 Directory.Delete("XA300", true);
 
             var a = CodeGeneratorArgs.Create<CodeGeneratorTest>("ValidEntity.yaml", "Data/ValidEntity.yaml").AddParameter("Directory", "XA300").AddParameter("AppName", "Zzz");
-            var c = CodeGenConsole.Create<CodeGenConsoleTest>(a);
+            var c = new CodeGenConsole(a);
             var r = await c.RunAsync();
             Assert.AreEqual(0, r);
 
@@ -118,7 +118,7 @@ namespace OnRamp.Test
             if (Directory.Exists("XA310"))
                 Directory.Delete("XA310", true);
 
-            var c = CodeGenConsole.Create<CodeGenConsoleTest>();
+            var c = new CodeGenConsole();
             var r = await c.RunAsync("-s ValidEntity.yaml -c Data/ValidEntity.yaml -p Directory=XA310 -p AppName=Zzz -a \"OnRamp.Test, Version=1.2.3.0, Culture=neutral, PublicKeyToken=null\"");
             Assert.AreEqual(0, r);
 
@@ -144,7 +144,7 @@ namespace OnRamp.Test
             if (Directory.Exists("XA400"))
                 Directory.Delete("XA400", true);
 
-            var c = CodeGenConsole.Create<CodeGenConsoleTest>();
+            var c = new CodeGenConsole();
             var r = await c.RunAsync("-s ValidEntity.yaml -c Data/ValidEntity.yaml -enc -p Directory=XA400 -p AppName=Zzz -a \"OnRamp.Test, Version=1.2.3.0, Culture=neutral, PublicKeyToken=null\"");
             Assert.AreEqual(3, r);
         }
@@ -230,7 +230,7 @@ namespace OnRamp.Test
             var ed = Environment.CurrentDirectory;
             Assert.IsTrue(ed.Contains(Path.Combine("bin", "debug"), StringComparison.InvariantCultureIgnoreCase) || ed.Contains(Path.Combine("bin", "release"), StringComparison.InvariantCultureIgnoreCase));
 
-            ed = CodeGenConsoleBase.GetBaseExeDirectory();
+            ed = CodeGenConsole.GetBaseExeDirectory();
             Assert.IsFalse(ed.Contains(Path.Combine("bin", "debug"), StringComparison.InvariantCultureIgnoreCase) || ed.Contains(Path.Combine("bin", "release"), StringComparison.InvariantCultureIgnoreCase));
         }